File reorg to separate `legacy` and `optimized` merkle implementation
